fix: revision 3

+++ src/Applications/use_case/thread-comments/DeleteThreadCommentUseCase.js
@@ -7,6 +7,11 @@ class DeleteThreadCommentUseCase {
 
   async execute (useCasePayload) {
     const payload = new DeleteThreadComment(useCasePayload)
+    await this._threadCommentRepository.checkAvailability(payload.id)
+    await this._threadCommentRepository.checkAccess({
+      commentId: payload.id,
+      userId: payload.userId
+    })
     return this._threadCommentRepository.deleteById(payload)
   }
 }
